chore: switch from controllers to minimal APIs
